+++ BlueprintCore/articles/changes.md
@@ -5,6 +5,7 @@
 * Added `SetLocalizedDuration` utility to `AbilityConfigurator` for easy reference of standard duration strings
 * Adding a spell to a mod spell list will no longer cause an NPE if the mod is not present
     * An error is still logged in BPCore using the Verbose channel
+* Fixed a bug causing validation to always fail for some components
 
 ## v2.6.2 Release